Refactor blog section to stacks, update navigation and metadata, and add new posts
