+++ app/bio.mdx
@@ -1,5 +1,5 @@
 Hi, welcome to my personal website. I am a senior studying mathematics at Stanford University, and also a coterminal Master's student in theoretical computer science. I'm interested broadly in geometric subjects, although more specifically I like studying algebraic geometry.
 
-I have a couple of **blog posts** that I write, which are mainly on mathematics that I find interesting but for which I don't have a full set of coherent notes. Feel free to check them out below.
+I have a blog on pure math called **The Stacks** (named after the stacks project), which is mainly on mathematics that I find interesting but for which I don't have a full set of coherent notes. Feel free to check this out below.
 
-My **projects** page contains some papers I have authored or co-authored, as well as various notes from classes I have taken while at Stanford. 
+My **Projects** page contains some papers I have authored or co-authored, as well as various notes from classes I have taken while at Stanford.
