+++ app/components/mdx.tsx
@@ -6,7 +6,7 @@ import React from 'react'
 import { compileMDX } from 'next-mdx-remote/rsc'
 import remarkMath from 'remark-math'
 import rehypeKatex from 'rehype-katex'
-import { generateMacros } from 'app/blog/utils'
+import { generateMacros } from 'app/stacks/utils'
 
 const katexCSS = '\n <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/katex@0.16.11/dist/katex.min.css" integrity="sha384-nB0miv6/jRmo5UMMR1wu3Gz6NLsoTkbqJghGIsx//Rlm+ZU03BU6SQNC66uf4l5+" crossOrigin="anonymous"/>'
 const options: KatexOptions = {
