+++ app/sitemap.ts
@@ -1,4 +1,4 @@
-import { getBlogPosts } from 'app/blog/utils'
+import { getBlogPosts } from 'app/stacks/utils'
 
 export const baseUrl = 'https://portfolio-blog-starter.vercel.app'
