+++ app/stacks/posts/affine-grassmannian-1.mdx
