+++ app/stacks/utils.ts
@@ -127,7 +127,7 @@ function getMDXData(dir) {
 }
 
 export function getBlogPosts() {
-  return getMDXData(path.join(process.cwd(), 'app', 'blog', 'posts'))
+  return getMDXData(path.join(process.cwd(), 'app', 'stacks', 'posts'))
 }
 
 export function formatDate(date: string, includeRelative = false) {
